Refactor sink/source for distributed for new grid msgsing
