+++ planner/gridtasks/sink_test.go
@@ -1,0 +1,9 @@
+package gridtasks
+
+import (
+	"testing"
+)
+
+func TestSink(t *testing.T) {
+
+}
